update config time storage to float

+++ Assets/Scripts/Libraries/Config.cs
@@ -33,7 +33,7 @@ namespace Config
         //TODO may need to have a levelData class to track with level select, or use configData?
         //TODO - totalLevelsCompleted can be gotten with .size() on LevelsCompleted - remove?
         public int TotalLevelsCompleted = 0;
-        public Dictionary<string, int> LevelBestTimes = new Dictionary<string, int>();
+        public Dictionary<string, float> LevelBestTimes = new Dictionary<string, float>();
         public Dictionary<string, int> LevelBestScores = new Dictionary<string, int>();
         public Dictionary<string, int> LevelBestActions = new Dictionary<string, int>();
 
@@ -47,7 +47,7 @@ namespace Config
         // constructor
         public ConfigData(
             int newTotalLevelsCompleted = 0,
-            Dictionary<string, int> newLevelBestTimes = null,
+            Dictionary<string, float> newLevelBestTimes = null,
             Dictionary<string, int> newLevelBestScores = null,
             Dictionary<string, int> newLevelBestActions = null,
             Dictionary<string, bool> newLevelsCompleted = null,
@@ -56,7 +56,7 @@ namespace Config
             )
         {
             TotalLevelsCompleted = newTotalLevelsCompleted;
-            LevelBestTimes = newLevelBestTimes ?? new Dictionary<string, int>();
+            LevelBestTimes = newLevelBestTimes ?? new Dictionary<string, float>();
             LevelBestScores = newLevelBestScores ?? new Dictionary<string, int>();
             LevelBestActions = newLevelBestActions ?? new Dictionary<string, int>();
             LevelsCompleted = newLevelsCompleted ?? new Dictionary<string, bool>();
